use uuids for album and artist routes
Previously the artist and album page routes attempted to pass artist and album names as params in the url. This caused all sorts of issues with weird characters in artist and album names. Stipping those weird characters out means that lastfm search results stopped working correctly. Yesterday it was realised that musicBrainz ids could be used to perform lastfm API requests which work much better when passed around in the url