Contingency for when lastFm returns no mbid
LastFms data is completely unreliable. Missing track information, missing mbids etc is common. As Tuneify uses mbids as a mechanism to identify resources it causes issues when they are not returned from LastFm. This work puts a fallback in place. If there is no mbid then just use the resource name (i.e. the artist name or album name) as an identifier