Remove express server altogether
Upon investigating rendering the layout template from express it was realised that the express app has become needless. Since the universal rendering was removed all it was doing was rendering the html. This can be done in development mode using webpack dev server and in production the app will be rendered from the generated html served from s3